requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.3.2/constraints-3.9.txt"
apache-airflow==2.3.2
pandas==1.4.2
